Nettoyage de code+repare les redirect ajout/delete
